refactor: move move_willo_by_table and Direction into movement_table module
