+++ src/gameplay/mod.rs
@@ -10,32 +10,6 @@ pub fn xy_translation(coords: IVec2) -> Vec2 {
 }
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
-pub enum Direction {
-    Up,
-    Left,
-    Down,
-    Right,
-}
-
-pub const DIRECTION_ORDER: [Direction; 4] = [
-    Direction::Up,
-    Direction::Left,
-    Direction::Down,
-    Direction::Right,
-];
-
-impl From<Direction> for IVec2 {
-    fn from(direction: Direction) -> IVec2 {
-        match direction {
-            Direction::Up => IVec2::Y,
-            Direction::Left => IVec2::new(-1, 0),
-            Direction::Down => IVec2::new(0, -1),
-            Direction::Right => IVec2::X,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
 pub struct DeathEvent {
     pub willo_entity: Entity,
 }
