+++ src/movement_table.rs
@@ -1,5 +1,10 @@
-use crate::{gameplay::components::*, *};
+use crate::{
+    gameplay::components::*,
+    willo::{WilloMovementEvent, WilloState},
+    *,
+};
 use bevy::prelude::*;
+use iyes_loopless::prelude::*;
 
 pub struct MovementTablePlugin;
 
@@ -10,7 +15,50 @@ impl Plugin for MovementTablePlugin {
                 .run_in_state(GameState::Gameplay)
                 .before(SystemLabels::Input),
         )
+        .add_system(
+            move_willo_by_table
+                .run_in_state(GameState::Gameplay)
+                .after(SystemLabels::MoveTableUpdate)
+                .after(history::FlushHistoryCommands),
+        )
         .register_ldtk_entity::<MovementTableBundle>("Table");
+    }
+}
+
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
+pub enum Direction {
+    Up,
+    Left,
+    Down,
+    Right,
+}
+
+pub const DIRECTION_ORDER: [Direction; 4] = [
+    Direction::Up,
+    Direction::Left,
+    Direction::Down,
+    Direction::Right,
+];
+
+impl From<Direction> for IVec2 {
+    fn from(direction: Direction) -> IVec2 {
+        match direction {
+            Direction::Up => IVec2::Y,
+            Direction::Left => IVec2::new(-1, 0),
+            Direction::Down => IVec2::new(0, -1),
+            Direction::Right => IVec2::X,
+        }
+    }
+}
+
+const MOVEMENT_SECONDS: f32 = 0.14;
+
+#[derive(Clone, Debug, Component)]
+pub struct MovementTimer(Timer);
+
+impl Default for MovementTimer {
+    fn default() -> MovementTimer {
+        MovementTimer(Timer::from_seconds(MOVEMENT_SECONDS, false))
     }
 }
 
@@ -42,6 +90,49 @@ fn movement_table_update(
             if (0..4).contains(&x_index) && (0..4).contains(&y_index) {
                 // key block is in table
                 table.table[y_index as usize][x_index as usize] = Some(input_block.key_code);
+            }
+        }
+    }
+}
+
+fn move_willo_by_table(
+    table_query: Query<&MovementTable>,
+    mut willo_query: Query<(&mut MovementTimer, &mut WilloState)>,
+    mut movement_writer: EventWriter<WilloMovementEvent>,
+    time: Res<Time>,
+) {
+    for table in table_query.iter() {
+        if let Ok((mut timer, mut willo)) = willo_query.get_single_mut() {
+            timer.0.tick(time.delta());
+
+            if timer.0.finished() {
+                match *willo {
+                    WilloState::RankMove(key) => {
+                        for (i, rank) in table.table.iter().enumerate() {
+                            if rank.contains(&Some(key)) {
+                                movement_writer.send(WilloMovementEvent {
+                                    direction: DIRECTION_ORDER[i],
+                                });
+                            }
+                        }
+                        *willo = WilloState::FileMove(key);
+                        timer.0.reset();
+                    }
+                    WilloState::FileMove(key) => {
+                        for rank in table.table.iter() {
+                            for (i, cell) in rank.iter().enumerate() {
+                                if *cell == Some(key) {
+                                    movement_writer.send(WilloMovementEvent {
+                                        direction: DIRECTION_ORDER[i],
+                                    });
+                                }
+                            }
+                        }
+                        *willo = WilloState::Waiting;
+                        timer.0.reset();
+                    }
+                    _ => {}
+                }
             }
         }
     }
